requirements.txt
importlib-metadata==1.7.0
Pillow==7.2.0
py7zr==0.9.5
pycryptodome==3.9.8
rarfile==4.0
texttable==1.6.2
zipp==3.1.0
